Add files via upload
Uploaded the dataset, it is a zip file that contains images for all 3 classes, the images are not split into train/test/val so they need to be split before training. The dataset contains a total of 2671 high-quality images, carefully curated and labeled to ensure accuracy and consistency. The images are in JPEG format and have a resolution of 224 x 224 x 3 pixels. Note that the images were sourced from Unsplash and are licensed under the Unsplash License, which allows free use of the images without attribution.